backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-storage==2.14.0
google-cloud-resource-manager==1.10.0
python-multipart==0.0.6
numpy>=1.21.0
pyarrow>=10.0.0
pandas>=1.5.0
pyiceberg[gcs]==0.6.0
fsspec>=2023.1.0
gcsfs>=2023.1.0
fastavro>=1.8.0

pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1
